Added priority queue

+++ src/main/java/ru/spbau/ir/books/Frontier.java
@@ -12,7 +12,7 @@ import java.util.*;
 public class Frontier {
     private Queue<Crawler.WebsiteAndUrl> queue;
 
-    private static Comparator<Crawler.WebsiteAndUrl> queueComparator = Comparator.comparingInt(o -> o.site.getDelayTime());
+    private static Comparator<Crawler.WebsiteAndUrl> queueComparator = Comparator.comparingDouble(o -> o.site.getDelayTime());
 
     Frontier(Path mainURLs, Path unhandled) {
         queue = new PriorityQueue<>(queueComparator);
